requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit>=1.30.0
torch>=2.0.0
transformers>=4.36.0
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0
python-multipart==0.0.6
pillow>=10.0.0
huggingface-hub>=0.23.0 